Finished applications API endpoint group

+++ app/Http/Controllers/ApplicationController.php
@@ -23,6 +23,7 @@ namespace App\Http\Controllers;
 
 use App\Application;
 use App\Events\ApplicationDeniedEvent;
+use App\Facades\JSON;
 use App\Http\Resources\ApplicationResource;
 use App\Notifications\ApplicationMoved;
 use App\Notifications\NewApplicant;
@@ -37,7 +38,7 @@ use Illuminate\Support\Facades\Log;
 class ApplicationController extends Controller
 {
 
-   
+
     private function canVote($votes): bool
     {
         $allvotes = collect([]);
@@ -59,11 +60,10 @@ class ApplicationController extends Controller
 
     public function showUserApp(Request $request, Application $application)
     {
-        if (!$request->wantsJson())
-        {
+        if (!$request->wantsJson()) {
             $this->authorize('view', $application);
 
-            if (! is_null($application)) {
+            if (!is_null($application)) {
                 return view('dashboard.user.viewapp')
                     ->with(
                         [
@@ -72,7 +72,7 @@ class ApplicationController extends Controller
                             'structuredResponses' => json_decode($application->response->responseData, true),
                             'formStructure' => $application->response->form,
                             'vacancy' => $application->response->vacancy,
-                            'canVote'  => $this->canVote($application->votes),
+                            'canVote' => $this->canVote($application->votes),
                         ]
                     );
             } else {
@@ -92,8 +92,7 @@ class ApplicationController extends Controller
 
     public function showAllApps(Request $request)
     {
-        if (!$request->wantsJson())
-        {
+        if (!$request->wantsJson()) {
             $this->authorize('viewAny', Application::class);
 
             return view('dashboard.appmanagement.all')
@@ -115,7 +114,7 @@ class ApplicationController extends Controller
 
         $firstVacancy = $vacancyWithForm->first();
 
-        if (! $vacancyWithForm->isEmpty() && $firstVacancy->vacancyCount !== 0 && $firstVacancy->vacancyStatus == 'OPEN') {
+        if (!$vacancyWithForm->isEmpty() && $firstVacancy->vacancyCount !== 0 && $firstVacancy->vacancyStatus == 'OPEN') {
             return view('dashboard.application-rendering.apply')
                 ->with([
                     'vacancy' => $vacancyWithForm->first(),
@@ -130,35 +129,67 @@ class ApplicationController extends Controller
     {
         $vacancy = Vacancy::with('forms')->where('vacancySlug', $vacancySlug)->get();
 
-        if ($vacancy->first()->vacancyCount == 0 || $vacancy->first()->vacancyStatus !== 'OPEN') {
-            $request->session()->flash('error', 'This application is unavailable.');
+        if ($vacancy->isEmpty()) {
+            if (!$request->wantsJson()) {
+                return redirect()
+                    ->back()
+                    ->with('error', 'This vacancy doesn\'t exist; Please use the proper buttons to apply to one.');
+            }
+            return JSON::setResponseType('error')
+                ->setStatus('error')
+                ->setMessage('Can\'t apply to non-existent application!')
+                ->setCode(404)
+                ->build();
+        }
 
-            return redirect()->back();
+        if ($vacancy->first()->vacancyCount == 0 || $vacancy->first()->vacancyStatus !== 'OPEN') {
+
+            if ($request->wantsJson()) {
+                return JSON::setResponseType('error')
+                    ->setStatus('error')
+                    ->setMessage('This application is unavailable.')
+                    ->setCode(404)
+                    ->build();
+            }
+
+            return redirect()
+                ->back()
+                ->with('error', 'This application is unavailable');
         }
 
         Log::info('Processing new application!');
 
         $formStructure = json_decode($vacancy->first()->forms->formStructure, true);
-        $responseValidation = ContextAwareValidator::getResponseValidator($request->all(), $formStructure);
+        $responseValidation = ($request->wantsJson()) ? ContextAwareValidator::getResponseValidator($request->json('payload'), $formStructure) : ContextAwareValidator::getResponseValidator($request->all(), $formStructure);
+        $applicant = ($request->wantsJson()) ? User::findOrFail($request->json('metadata.submittingUserID')) : Auth::user();
+
+        // API users may specify ID 1 for an anonymous application, but they'll have to submit contact details for it to become active.
+        // User ID 1 is exempt from application rate limits
 
         Log::info('Built response & validator structure!');
 
-        if (! $responseValidation->get('validator')->fails()) {
+        if (!$responseValidation->get('validator')->fails()) {
             $response = Response::create([
                 'responseFormID' => $vacancy->first()->forms->id,
                 'associatedVacancyID' => $vacancy->first()->id, // Since a form can be used by multiple vacancies, we can only know which specific vacancy this response ties to by using a vacancy ID
                 'responseData' => $responseValidation->get('responseStructure'),
             ]);
 
-            Log::info('Registered form response for user '.Auth::user()->name.' for vacancy '.$vacancy->first()->vacancyName);
+            Log::info('Registered form response!', [
+                'applicant' => $applicant->name,
+                'vacancy' => $vacancy->first()->vacancyName
+            ]);
 
             $application = Application::create([
-                'applicantUserID' => Auth::user()->id,
+                'applicantUserID' => $applicant->id,
                 'applicantFormResponseID' => $response->id,
                 'applicationStatus' => 'STAGE_SUBMITTED',
             ]);
 
-            Log::info('Submitted application for user '.Auth::user()->name.' with response ID'.$response->id);
+            Log::info('Submitted an application!', [
+                'responseID' => $response->id,
+                'applicant' => $applicant->name
+            ]);
 
             foreach (User::all() as $user) {
                 if ($user->hasRole('admin')) {
@@ -166,37 +197,73 @@ class ApplicationController extends Controller
                 }
             }
 
-            $request->session()->flash('success', 'Thank you for your application! It will be reviewed as soon as possible.');
+            if ($request->wantsJson()) {
+                return JSON::setResponseType('success')
+                    ->setStatus('accepted')
+                    ->setMessage('Application submitted successfully. Please refer to the docs to add contact info if your submission was anonymous.')
+                    ->setCode(201)
+                    ->setAdditional([
+                        'links' => [
+                            'application-web' => route('showUserApp', ['application' => $application->id])
+                        ]
+                    ])->build();
+            }
 
-            return redirect()->to(route('showUserApps'));
-        } else {
-            Log::warning('Application form for '.Auth::user()->name.' contained errors, resetting!');
-            $request->session()->flash('error', 'There are one or more errors in your application. Please make sure none of your fields are empty, since they are all required.');
+            $request->session()->flash('success', 'Thank you for your application! It will be reviewed as soon as possible.');
+            return redirect(route('showUserApps'));
+
+        } elseif ($request->wantsJson()) {
+            return JSON::setResponseType('error')
+                ->setStatus('validation_error')
+                ->setMessage('There are one or more errors in this application. Please make sure all fields are present in "payload" according to this application\'s respective form structure; They\'re always required.')
+                ->setCode(400)
+                ->build();
         }
 
-        return redirect()->back();
+        Log::warning('Application form for ' . $applicant->name . ' contained errors, resetting!');
+
+        return redirect()
+            ->back()
+            ->with('error', 'There are one or more errors in your application. Please make sure none of your fields are empty, since they are all required.');
     }
 
     public function updateApplicationStatus(Request $request, Application $application, $newStatus)
     {
-        $this->authorize('update', Application::class);
+        $messageIsError = false;
+
+        if (!$request->wantsJson())
+            $this->authorize('update', Application::class);
+
 
         switch ($newStatus) {
             case 'deny':
 
                 event(new ApplicationDeniedEvent($application));
+                $message = "Application denied successfully.";
+
                 break;
 
             case 'interview':
-                Log::info('User '.Auth::user()->name.' has moved application ID '.$application->id.'to interview stage');
-                $request->session()->flash('success', 'Application moved to interview stage! (:');
-                $application->setStatus('STAGE_INTERVIEW');
+                Log::info(' Moved application ID ' . $application->id . 'to interview stage!');
+                $message = 'Application moved to interview stage! (:';
 
+                $application->setStatus('STAGE_INTERVIEW');
                 $application->user->notify(new ApplicationMoved());
+
                 break;
 
             default:
-                $request->session()->flash('error', 'There are no suitable statuses to update to. Do not mess with the URL.');
+                $message = "There are no suitable statuses to update to.";
+                $messageIsError = true;
+        }
+
+        if ($request->wantsJson())
+        {
+            return JSON::setResponseType(($messageIsError) ? 'error' : 'success')
+                ->setStatus(($messageIsError) ? 'error' : 'success')
+                ->setMessage($message)
+                ->setCode(($messageIsError) ? 400 : 200)
+                ->build();
         }
 
         return redirect()->back();
@@ -204,11 +271,22 @@ class ApplicationController extends Controller
 
     public function delete(Request $request, Application $application)
     {
-        $this->authorize('delete', $application);
-        $application->delete(); // observers will run, cleaning it up
+        if (!$request->wantsJson()) {
+            $this->authorize('delete', $application);
+            $application->delete(); // observers will run, cleaning it up
 
-        $request->session()->flash('success', 'Application deleted. Comments, appointments and responses have also been deleted.');
+            return redirect()
+                ->back()
+                ->with('success', 'Application deleted. Comments, appointments and responses have also been deleted.');
+        }
 
-        return redirect()->back();
+
+        $application->delete();
+
+        return JSON::setResponseType('success')
+            ->setStatus('deleted')
+            ->setMessage('Application deleted. Relationships were also nuked.')
+            ->setCode(200)
+            ->build();
     }
 }
